Dept filters

+++ src/main/java/com/getkhaki/api/bff/persistence/models/views/TimeBlockSummaryView.java
@@ -9,7 +9,7 @@ public interface TimeBlockSummaryView {
     Long getMeetingCount();
     Long getNumEmployees();
     Long getNumWorkdays();
-    Long getTotalInternalMeetingAttendees();
-    Long getTotalMeetingAttendees();
+    Integer getTotalInternalMeetingAttendees();
+    Integer getTotalMeetingAttendees();
     Long getMeetingLengthSeconds();
 }
